requirements.txt
azure-functions
azure-cognitiveservices-vision-face
python-dotenv
Pillow
requests